+++ src/installers/registry.ts
@@ -2,9 +2,11 @@ import { BepInExInstaller } from "./BepInExInstaller";
 import { GodotMLInstaller } from "./GodotMLInstaller";
 import { MelonLoaderInstaller } from "./MelonLoaderInstaller";
 import { PackageInstaller } from "./PackageInstaller";
+import { InstallRuleInstaller } from "src/installers/InstallRuleInstaller";
 
 
 const _PackageInstallers = {
+    // "legacy": new InstallRuleInstaller(),  // TODO: Enable
     "bepinex": new BepInExInstaller(),
     "godotml": new GodotMLInstaller(),
     "melonloader": new MelonLoaderInstaller(),
